confirmation popup

+++ app/components/expenses/ExpenseListItem.jsx
@@ -4,6 +4,10 @@ function ExpenseListItem({ id, title, amount }) {
   const fetcher = useFetcher()
 
   function deleteExpenseItemHandler() {
+    const proceed = confirm('Are you sure you want to delete?')
+
+    if (!proceed) return
+    
     fetcher.submit(null, {
       method: 'delete',
       action: `/expenses/${id}`
